テロップ実装、テクスチャ差し替え

+++ project/source/Viewer/GameScene/GameViewer/TimerViewer.h
@@ -28,7 +28,7 @@ public:
 	void Draw(void);
 
 	//ƒpƒ‰ƒ[ƒ^‚ğó‚¯‚Æ‚é” 
-	int parameterBox;
+	float parameterBox;
 
 private:
 	BaseViewerDrawer *bg;
